server/requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
boto3>=1.24.0
python-dotenv>=0.20.0
numpy>=1.24.0
scikit-learn>=1.2.2
Pillow>=9.5.0
sqlalchemy>=1.4.46
psycopg2-binary>=2.9.5
requests>=2.28.2
modal>=0.51.3
apscheduler>=3.10.1